🛠️ refactor: 상품 목록 조회 리펙토링
- 구현계층 삭제
- keyword 빈 값으로 받을 수 있게 변경
- 서비스 레이어 분리

🔗 Resolves: #294
