--- backend/JiShop/src/main/java/com/jishop/product/implementation/ProductQueryHelper.java
@@ -1,12 +1,0 @@
-package com.jishop.product.implementation;
-
-import com.jishop.product.domain.QProduct;
-import com.jishop.product.dto.request.ProductRequest;
-import com.jishop.reviewproduct.domain.QReviewProduct;
-import com.querydsl.core.BooleanBuilder;
-
-public interface ProductQueryHelper {
-
-    BooleanBuilder findProductsByCondition(final ProductRequest request,
-                                           final QProduct product, final QReviewProduct reviewProduct);
-}
