+++ backend/JiShop/src/main/java/com/jishop/product/repository/ProductRepositoryQueryDsl.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ProductRepositoryQueryDsl {
 
-    List<Product> findProductsByCondition(final ProductRequest request, final int page, final int size);
+    List<Product> findProductsByCondition(final ProductRequest request, final int page, final int size, final List<Long> categoryIds);
 
-    long countProductsByCondition(final ProductRequest request);
+    long countProductsByCondition(final ProductRequest request,  final List<Long> categoryIds);
 }
